feat(ux): add confirmation step for deleting experience items
Extract ExperienceItem to its own component and implement a confirmation UI
when deleting an item to prevent accidental data loss. This improves
accessibility and safety.

- Extract `ExperienceItem` from `pages/Editor.tsx` to `components/ExperienceItem.tsx`
- Add `isDeleting` state to `ExperienceItem`
- Show check/cancel buttons instead of immediate delete
- Add unit tests for the new component and interaction
